PaaS-316 - Add Package Install Option

+++ recipes/default.rb
@@ -17,42 +17,49 @@
 # under the License.
 #
 
-bash "install-golang" do
-  cwd Chef::Config[:file_cache_path]
-  code <<-EOH
+if node['go']['package_install']
+  package 'install golang' do
+    package_name 'golang'
+    version node['go']['package_version']
+  end
+else
+  bash 'install-golang' do
+    cwd Chef::Config[:file_cache_path]
+    code <<-EOH
     rm -rf go
     rm -rf #{node['go']['install_dir']}/go
-    tar -C #{node['go']['install_dir']} -xzf #{node["go"]["filename"]}
-  EOH
-  action :nothing
+    tar -C #{node['go']['install_dir']} -xzf #{node['go']['filename']}
+    EOH
+    action :nothing
+  end
+
+  remote_file File.join(Chef::Config[:file_cache_path], node['go']['filename']) do
+    source node['go']['url']
+    owner 'root'
+    mode 0644
+    notifies :run, 'bash[install-golang]', :immediately
+    not_if "#{node['go']['install_dir']}/go/bin/go version | grep \"go#{node['go']['version']} \""
+  end
+
+  directory node['go']['gopath'] do
+    action :create
+    recursive true
+    owner node['go']['owner']
+    group node['go']['group']
+    mode node['go']['mode']
+  end
+
+  directory node['go']['gobin'] do
+    action :create
+    recursive true
+    owner node['go']['owner']
+    group node['go']['group']
+    mode node['go']['mode']
+  end
 end
 
-remote_file File.join(Chef::Config[:file_cache_path], node['go']['filename']) do
-  source node['go']['url']
-  owner 'root'
-  mode 0644
-  notifies :run, 'bash[install-golang]', :immediately
-  not_if "#{node['go']['install_dir']}/go/bin/go version | grep \"go#{node['go']['version']} \""
-end
-
-directory node['go']['gopath'] do
-  action :create
-  recursive true
-  owner node['go']['owner']
-  group node['go']['group']
-  mode node['go']['mode']
-end
-
-directory node['go']['gobin'] do
-  action :create
-  recursive true
-  owner node['go']['owner']
-  group node['go']['group']
-  mode node['go']['mode']
-end
-
-template "/etc/profile.d/golang.sh" do
-  source "golang.sh.erb"
+template '/etc/profile.d/golang.sh' do
+  source 'golang.sh.erb'
   owner 'root'
   group 'root'
   mode 0755
